requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap-v5==1.0.6
django-crispy-forms==1.13.0
django-dob-widget==1.2.0
django-embed-video==1.4.0
django-heroku==0.3.1
django-registration==3.2
django-session-timeout==0.1.0
django-taggit==1.5.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
isodate==0.6.0
Pillow==8.4.0
psycopg2==2.9.1
python-decouple==3.5
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
